Move reward function to a new file
