+++ torchtune/datasets/_rl.py
@@ -5,13 +5,10 @@
 # LICENSE file in the root directory of this source tree.
 
 from typing import Any, Callable, Dict, Mapping, Optional, TypedDict
-from xml.etree import ElementTree as ET
 
-import torch
 from datasets import load_dataset
 from torch.utils.data import Dataset
 
-from torchtune import utils
 from torchtune.modules.tokenizers import ModelTokenizer
 from torchtune.modules.transforms import Transform
 
@@ -73,97 +70,3 @@ class RLDataset(Dataset):
         answer = transformed_sample["answer"]
 
         return {"tokens": q_tokens, "mask": mask, "answer": answer}
-
-
-def extract_tags(text: str) -> dict[str, list[str]]:
-    """
-    Parse XML-like tags from text. Returns a dictionary with keys 'think' and 'answer'.
-    The values are lists of strings, with each string being the content of a tag.
-    """
-    xml_string = f"<root>{text}</root>"
-    root = ET.fromstring(xml_string)
-
-    return {
-        "think": [
-            elem.text if elem.text is not None else "" for elem in root.findall("think")
-        ],
-        "answer": [
-            elem.text if elem.text is not None else ""
-            for elem in root.findall("answer")
-        ],
-    }
-
-
-def shaped_correctness_reward(
-    question: str, answer: str, completion: str
-) -> tuple[float, float]:
-    """
-    Reward function for verifiable rewards with some mild shaping.
-    """
-    question_chars = len(question)
-    only_completion = completion[question_chars:]
-
-    reward = 0
-    success = 0
-
-    try:
-        tags = extract_tags(
-            "<think>" + only_completion.replace("<<", "").replace(">>", "")
-        )
-    except ET.ParseError:
-        tags = {"think": [], "answer": []}
-
-    if len(tags["answer"]) == 1:
-        reward += 5.0
-
-    if len(tags["think"]) == 1:
-        reward += 5.0
-
-    if any(attempt == answer for attempt in tags["answer"]):
-        # One of the answer tags has the right answer
-        reward += 20.0
-
-    if any((answer in attempt) for attempt in tags["answer"]):
-        # One of the answer tags contains the right answer (might be e.g. $20 instead of 20)
-        reward += 10.0
-
-    # total_think_length = sum(len(c) + len("<think></think>") for c in tags["think"])
-    # total_answer_length = sum(len(a) + len("<answer></answer>") for a in tags["answer"])
-    # total_extra_length = len(only_completion) - total_think_length - total_answer_length
-
-    # reward -= 0.1 * total_extra_length
-
-    if len(tags["answer"]) > 0 and tags["answer"][-1] == answer:
-        reward = 100.0
-        success = 1
-
-    _, rank = utils.get_world_size_and_rank()
-    if rank == 0:
-        print()
-        print(f"{question=}\n{answer=}\n{reward=}\n{only_completion=}")
-        print()
-
-    return reward, success
-
-
-def batch_shaped_correctness_reward(
-    tokenizer: ModelTokenizer, completions: torch.Tensor, answers: list[str]
-) -> [torch.Tensor, torch.Tensor]:
-    """Utility function to apply the shaped reward function to a GRPO-style batch of completions."""
-
-    batch_size, grpo_size, *_ = completions.shape
-    rewards = torch.zeros(batch_size, grpo_size, dtype=torch.float32)
-    successes = torch.zeros(batch_size, grpo_size, dtype=torch.float32)
-    # completions :: [B, G, L]
-    for b in range(batch_size):
-        for g in range(grpo_size):
-            text_completion = tokenizer.decode(
-                completions[b, g].tolist()
-            )  # skips special tokens, stops at eos
-            reward, success = shaped_correctness_reward(
-                question="", answer=answers[b], completion=text_completion
-            )
-            rewards[b, g] = reward
-            successes[b, g] = success
-
-    return rewards, successes
